+++ src/main/kotlin/com/githubyss/mobile/common/ui/banner/classical/BannerCarouselView.kt
@@ -25,15 +25,15 @@ import java.lang.ref.WeakReference
  * @createdTime 2021/04/06 10:52:05
  */
 class BannerCarouselView : FrameLayout {
-    
+
     /** ****************************** Properties ****************************** */
-    
+
     companion object {
         val TAG: String = BannerCarouselView::class.java.simpleName
         private const val SWITCH_VIEW_PAGER = 1
         private const val TIME_PERIOD = 5000
     }
-    
+
     private var indicatorWidth = 0.0f
     private var indicatorHeight = 0.0f
     private var indicatorMargin = 0
@@ -43,28 +43,28 @@ class BannerCarouselView : FrameLayout {
     private var actualCount = 0
     private var viewPager: BannerViewPager? = null
     private var indicator: BannerLineIndicator? = null
-    
-    
+
+
     /** ****************************** Constructors ****************************** */
-    
+
     constructor(context: Context) : this(context, null, 0)
     constructor(context: Context, attrs: AttributeSet?) : this(context, attrs, 0)
     constructor(context: Context, attrs: AttributeSet?, defStyleAttr: Int) : super(context, attrs, defStyleAttr) {
         init()
         initView(context)
     }
-    
-    
+
+
     /** ****************************** Override ****************************** */
-    
-    
+
+
     /** ****************************** Functions ****************************** */
-    
+
     fun setAdapter(adapter: PagerAdapter?) {
         if (isInit) {
             return
         }
-        
+
         isInit = true
         pageAdapter = adapter
         pageAdapter?.registerDataSetObserver(dataSetObserver)
@@ -74,7 +74,8 @@ class BannerCarouselView : FrameLayout {
         if (pageAdapter?.count == 1) {
             viewPager.setCurrentItem(0, false)
             dotIndicator.visibility = View.GONE
-        } else {
+        }
+        else {
             viewPager.setCurrentItem(1, false)
             dotIndicator.visibility = View.VISIBLE
             val params = dotIndicator.layoutParams
@@ -84,11 +85,12 @@ class BannerCarouselView : FrameLayout {
         dotIndicator.bindSlideWithViewPager(viewPager)
         actualCount = if ((pageAdapter?.count ?: return) > 2) {
             pageAdapter?.count ?: return
-        } else {
+        }
+        else {
             1
         }
     }
-    
+
     fun startCarouse() {
         if (pageAdapter == null || (pageAdapter?.count ?: return) < 2) {
             // 数量不够轮播时不进行轮播操作
@@ -98,24 +100,24 @@ class BannerCarouselView : FrameLayout {
         handler?.removeMessages(SWITCH_VIEW_PAGER)
         handler?.postDelayed(autoCarouselRunnable, TIME_PERIOD.toLong())
     }
-    
+
     fun stopCarouse() {
         handler?.removeCallbacks(autoCarouselRunnable)
         handler?.removeMessages(SWITCH_VIEW_PAGER)
     }
-    
+
     private fun init() {
         handler = MyHandler(this)
         indicatorWidth = dp2Px(9.0f).toFloat()
         indicatorHeight = dp2Px(1.5f).toFloat()
-        indicatorMargin = dp2Px(5.0f)
+        indicatorMargin = dp2Px(5.0f).toInt()
     }
-    
+
     private fun initView(context: Context) {
         if (isInit) {
             return
         }
-        
+
         viewPager = BannerViewPager(context)
         addView(viewPager)
         val viewPagerParams = viewPager?.layoutParams as LayoutParams
@@ -123,7 +125,7 @@ class BannerCarouselView : FrameLayout {
         viewPagerParams.height = ViewGroup.LayoutParams.MATCH_PARENT
         viewPager?.layoutParams = viewPagerParams
         viewPager?.onViewPagerTouchListener = onViewPagerTouchListener
-        
+
         indicator = BannerLineIndicator(context)
         val indicatorParams = LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT)
         indicatorParams.bottomMargin = indicatorMargin
@@ -131,17 +133,18 @@ class BannerCarouselView : FrameLayout {
         indicatorParams.gravity = Gravity.BOTTOM or Gravity.CENTER_HORIZONTAL
         indicator?.layoutParams = indicatorParams
         addView(indicator)
-        
+
         viewPager?.addOnPageChangeListener(onPageChangeListener)
     }
-    
+
     private fun updateDataSource() {
         val viewPager = getChildAt(0) as ViewPager
         val dotIndicator = getChildAt(1) as BannerLineIndicator
         if (pageAdapter?.count == 1) {
             viewPager.setCurrentItem(0, false)
             dotIndicator.visibility = View.GONE
-        } else {
+        }
+        else {
             viewPager.setCurrentItem(1, false)
             dotIndicator.visibility = View.VISIBLE
             val params = dotIndicator.layoutParams
@@ -152,17 +155,18 @@ class BannerCarouselView : FrameLayout {
         dotIndicator.updateViewPagerCount(viewPager)
         actualCount = if ((pageAdapter?.count ?: return) > 2) {
             pageAdapter?.count ?: return
-        } else {
+        }
+        else {
             1
         }
     }
-    
-    
+
+
     /** ****************************** Classes ****************************** */
-    
+
     internal class MyHandler(carouselView: BannerCarouselView) : Handler() {
         var weakReference: WeakReference<BannerCarouselView> = WeakReference(carouselView)
-        
+
         override fun handleMessage(msg: Message) {
             super.handleMessage(msg)
             when (msg.what) {
@@ -170,7 +174,8 @@ class BannerCarouselView : FrameLayout {
                     var viewPager: ViewPager? = null
                     try {
                         viewPager = weakReference.get()?.getChildAt(0) as ViewPager
-                    } catch (e: Exception) {
+                    }
+                    catch (e: Exception) {
                         logE(TAG, "first child must be ViewPager!")
                     }
                     if (viewPager != null && viewPager.childCount > 1) {
@@ -182,10 +187,10 @@ class BannerCarouselView : FrameLayout {
             }
         }
     }
-    
-    
+
+
     /** ****************************** Implementations ****************************** */
-    
+
     private val onPageChangeListener: OnPageChangeListener = object : OnPageChangeListener {
         override fun onPageScrolled(i: Int, v: Float, i1: Int) {}
         override fun onPageSelected(position: Int) {}
@@ -194,23 +199,23 @@ class BannerCarouselView : FrameLayout {
             }
         }
     }
-    
+
     private val onViewPagerTouchListener: BannerViewPager.OnViewPagerTouchListener = object : BannerViewPager.OnViewPagerTouchListener {
         override fun onTouchStart() {
             stopCarouse()
         }
-        
+
         override fun onTouchEnd() {
             startCarouse()
         }
     }
-    
+
     private val dataSetObserver: DataSetObserver = object : DataSetObserver() {
         override fun onChanged() {
             updateDataSource()
         }
     }
-    
+
     private val autoCarouselRunnable: Runnable = object : Runnable {
         override fun run() {
             handler?.sendEmptyMessage(SWITCH_VIEW_PAGER)
